enum Colors
Add enum Colors and other fixes
