+++ NewsToDay/Extensions/Colors.swift
@@ -1,0 +1,24 @@
+//
+//  Colors.swift
+//  NewsToDay
+//
+//  Created by Марк Райтман on 14.05.2023.
+//
+
+import Foundation
+
+enum Colors {
+    static let blackDarker = "Black Darker"
+    static let blackLight = "Black Light"
+    static let blackLighter = "Black Lighter"
+    static let blackPrimary = "Black Primary"
+    static let greyDarker = "Grey Darker"
+    static let greyLight = "Grey Light"
+    static let greyLighter  = "Grey Lighter"
+    static let greyPrimary  = "Grey Primary"
+    static let purpleDarker  = "Purple Darker"
+    static let purpleLight  = "Purple Light"
+    static let purpleLighter  = "Purple Lighter"
+    static let purplePrimary  = "Purple Primary"
+    static let white = "White"
+}
