--- NewsToDay/Resources/Resources.swift
@@ -1,28 +1,0 @@
-//
-//  Resources.swift
-//  NewsToDay
-//
-//  Created by Daniil Kulikovskiy on 10.05.2023.
-//
-
-import Foundation
-
-enum Resources {
-    
-    enum Colors {
-        static let blackDarker = "Black Darker"
-        static let blackLight = "Black Light"
-        static let blackLighter = "Black Lighter"
-        static let blackPrimary = "Black Primary"
-        static let greyDarker = "Grey Darker"
-        static let greyLight = "Grey Light"
-        static let greyLighter  = "Grey Lighter"
-        static let greyPrimary  = "Grey Primary"
-        static let purpleDarker  = "Purple Darker"
-        static let purpleLight  = "Purple Light"
-        static let purpleLighter  = "Purple Lighter"
-        static let purplePrimary  = "Purple Primary"
-        static let white = "White"
-    }
-    
-}
